requirements.txt
streamlit
requests
Pillow
python-magic-bin==0.4.14
